Add example test and update README accordingly
The testcases used in the CI Pipeline test
samples built with all possible frameworks.
During the example described in the README only
one sample variant is built.
Additionally, the build locations and linkers used
differ. Because of that a special test fitted
for the README example is added.

The pipeline is instructed to ignore the
new test, as it would fail during its execution,
due to the reasons named above.