Introduce `ARM_TARGET` property
Introduces a property named `ARM_TARGET` to specify the target board for
which the samples should be build. The property defaults to `STM32F407`.
Building samples for targets other than `STM32F407` or `STM32F4XX` is
restricted to CMSIS.
Amongst other things, corresponding strings have to be added to the
`ARM_TARGET_VALUES` list to expand support for new targets.