Add concurrency groups to pull requests
Cancels in-progress jobs for a current workflow (when pushing
additional commits to a branch that belongs to a PR). The docs suggest
to use the `github.ref`, but it seems that this also cancels workflows
on the main branch if more recent workflows for the main branch get
triggered.

References:
https://docs.github.com/en/actions/using-jobs/using-concurrency#example-only-cancel-in-progress-jobs-or-runs-for-the-current-workflow
https://docs.github.com/en/developers/webhooks-and-events/webhooks/webhook-events-and-payloads#pull_request